refactor: 메서드 순서 수정
